# Cache compiled templates for partials