@hugopeixoto 2015-04-29
Compiling a partial seems to depend on two things:its name and the indentation level. Maybe using that pair as the cache key would make some sense (to avoid the memory overhead of keeping the whole string around)