@evilstreak 2015-05-05
@hugopeixoto The problem with caching based on the name of a partial is that the content of the partial is not guaranteed to be the same from one use to the next. Here's a contrived example:

```
base.mustache:
{{#people}}
  {{> hello}}
{{/people}}

hello.mustache
Hello {{name}}
```

```ruby
class MyMustache < Mustache
  def partial(name)
    @counts ||= {}
    @counts[name] ||= 0

    count_string = "-- this partial has been included #{@counts[name]} times already\n"

    @counts[partial] += 1

    super + count_string
  end
end
```

That would output something like:

```
Hello Dom
-- this partial has been included 0 times already
Hello Hugo
-- this partial has been included 1 times already
Hello Ricardo
-- this partial has been included 2 times already
```

I have no way of knowing how many people are using the gem like that, but it would suck to break it for them. There was definitely concern about people using dynamically generated templates in #173, though I'm not sure if it was with a use case like this in mind.

The only way to be sure that the template we want is equivalent to one we've used before is to use the full string.