@hugopeixoto 2015-05-05
@evilstreak you're right, I forgot that we allowed overriding `#partial`.
I guess that in that dynamic example you'd also be using a large amount of memory, as each new template would produce a new entry in the cache table, but at least it would be wiped when the full render had finished. Eventually we could implement an LRU cache to prevent that.

The only thing that comes to mind that could be a problem (and I'm playing devil's advocate here) is if someone overrode `#render` in their mustache view, but that doesn't seem very plausible.