@evilstreak 2015-05-07
I considered overriding `#render` to do caching in our subclass but it's such an involved method it felt like it would be a really brittle option, likely to break with future versions of the gem.