@evilstreak 2015-05-15
Is this PR something you're broadly open to? It would be good to get a "yes, this feature is fine in principle but still needs a review to consider the approach taken and the code style" or a "no, this isn't a feature that should be in this library".