@hugopeixoto 2015-05-15
Personally I don't see anything wrong with the feature being included, but I'm not a maintainer. I'll try to ping @locks to review this.

What do you think about adding an LRU? Does it make sense? Sam Saffron has a gem called lru_redux (https://github.com/SamSaffron/lru_redux) which might be easy to add.