@evilstreak 2015-05-15
Given that the cache is only going to be hanging around for the lifetime of the `Mustache` object, I think going for a more advanced caching approach is overkill.

The only case where this naive caching could cause problems is where someone has a large number of different, large templates, and a mid-render garbage collection is necessary to stop the process running out of memory.

I don't think guarding against that case is worth the cost of including a dependency, especially since this library doesn't depend on any other gems at the moment.