@judofyr 2015-06-24
We could also make `template_for_partial` public API. Anyone who needs a special caching technique can override this method.